Fixed build for go1.20
