+++ internal/xslices/sort.go
@@ -7,14 +7,6 @@ import (
 	"slices"
 )
 
-func SortCopy[T any](in []T, cmp func(lhs, rhs T) int) (out []T) {
-	out = Clone(in)
-
-	Sort(out, cmp)
-
-	return out
-}
-
 func Sort[T any](in []T, cmp func(lhs, rhs T) int) {
 	slices.SortFunc(in, cmp)
 }
