+++ internal/xslices/sort_copy.go
@@ -1,0 +1,9 @@
+package xslices
+
+func SortCopy[T any](in []T, cmp func(lhs, rhs T) int) (out []T) {
+	out = Clone(in)
+
+	Sort(out, cmp)
+
+	return out
+}
